refactor maven build

+++ game/lwjgl/src/main/java/im/bci/newtonadv/platform/lwjgl/LwjglHelper.java
@@ -89,7 +89,7 @@ public class LwjglHelper {
                 methods = new Methods();
             }
             if (null != methods.getWidthMethod) {
-                return (int) methods.getWidthMethod.invoke(null);
+                return (int) (Integer) methods.getWidthMethod.invoke(null);
             }
         } catch (Exception e) {
         }
@@ -102,7 +102,7 @@ public class LwjglHelper {
                 methods = new Methods();
             }
             if (null != methods.getHeightMethod) {
-                return (int) methods.getHeightMethod.invoke(null);
+                return (int) (Integer)methods.getHeightMethod.invoke(null);
             }
         } catch (Exception e) {
         }
